refactor(agent): flatten driver module structure
- Remove drivers/runtime/ subdirectory; move all driver files to drivers/
- Inline AgentRuntime enum into agent/mod.rs (was agent/runtime.rs)
- Move command probe utilities to utils/cmd.rs
- Remove dead code: acp_adaptor_binary, binary_name
- Update all import paths across src/ and tests/
