+++ src/agent/drivers/claude.rs
@@ -9,8 +9,8 @@ use tokio::sync::mpsc;
 use tracing::{debug, trace, warn};
 
 use super::claude_headless::{self, HeadlessEvent};
-use crate::agent::drivers::{command_exists, run_command};
 use crate::agent::AgentRuntime;
+use crate::utils::cmd::{command_exists, run_command};
 
 use super::*;
