+++ src/agent/drivers/codex.rs
@@ -15,8 +15,8 @@ use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::sync::mpsc;
 use tracing::{debug, trace, warn};
 
-use crate::agent::drivers::{command_exists, run_command};
 use crate::agent::AgentRuntime;
+use crate::utils::cmd::{command_exists, run_command};
 
 use super::codex_app_server::{self, AppServerEvent, AppServerPhase, ItemEvent, TurnStatus};
 use super::*;
