+++ src/agent/mod.rs
@@ -4,11 +4,42 @@ pub mod drivers;
 mod event_forwarder;
 pub mod lifecycle;
 pub mod manager;
-pub mod runtime;
 pub mod runtime_status;
 pub mod templates;
 pub mod trace;
 pub mod workspace;
 
 pub use lifecycle::AgentLifecycle;
-pub use runtime::AgentRuntime;
+
+use serde::{Deserialize, Serialize};
+
+/// Supported local agent runtimes.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum AgentRuntime {
+    Claude,
+    Codex,
+    Kimi,
+    Opencode,
+}
+
+impl AgentRuntime {
+    pub const fn as_str(self) -> &'static str {
+        match self {
+            Self::Claude => "claude",
+            Self::Codex => "codex",
+            Self::Kimi => "kimi",
+            Self::Opencode => "opencode",
+        }
+    }
+
+    pub fn parse(value: &str) -> Option<Self> {
+        match value {
+            "claude" => Some(Self::Claude),
+            "codex" => Some(Self::Codex),
+            "kimi" => Some(Self::Kimi),
+            "opencode" => Some(Self::Opencode),
+            _ => None,
+        }
+    }
+}
